+++ app/_layout.tsx
@@ -2,12 +2,18 @@ import { Provider } from "react-redux";
 import store from "@/store";
 
 import { Stack } from "expo-router";
+import TopActionsBar from "@/components/TopActionsBar";
 
 const RootLayout = () => {
   return (
     <Provider store={store}>
       <Stack>
-        <Stack.Screen name="index" options={{ headerTitle: "" }} />
+        <Stack.Screen
+          name="index"
+          options={{
+            header: () => null,
+          }}
+        />
       </Stack>
     </Provider>
   );